adding DTOs

+++ src/main/java/com/oussama/eshop/services/impl/CartServiceImpl.java
@@ -1,46 +1,63 @@
 package com.oussama.eshop.services.impl;
 
+import com.oussama.eshop.domain.dto.CartDto;
+import com.oussama.eshop.domain.dto.ProductDto;
 import com.oussama.eshop.domain.entities.Cart;
+import com.oussama.eshop.domain.entities.Product;
+import com.oussama.eshop.mappers.Mapper;
 import com.oussama.eshop.repositories.CartRepository;
 import com.oussama.eshop.services.CartService;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class CartServiceImpl implements CartService {
 
 
     private final CartRepository cartRepository;
+    private final Mapper<Cart,CartDto> mapper;
 
-    public CartServiceImpl(CartRepository cartRepository) {
+    public CartServiceImpl(CartRepository cartRepository, Mapper<Cart, CartDto> mapper) {
         this.cartRepository = cartRepository;
+        this.mapper = mapper;
     }
 
     @Override
-    public Cart save(Cart cart) {
-        return cartRepository.save(cart);
+    public CartDto save(CartDto cart) {
+        return mapper.mapTo(mapper.mapFrom(cart));
     }
 
     @Override
-    public List<Cart> findAll() {
-        return cartRepository.findAll();
+    public List<CartDto> findAll() {
+        return cartRepository.findAll().stream().map(mapper::mapTo).toList();
     }
 
     @Override
-    public Optional<Cart> findOne(Integer id) {
-        return cartRepository.findById(id);
+    public CartDto findOne(Integer id) {
+        Cart cart = cartRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Cart not found with id: " + id));
+        return mapper.mapTo(cart);
     }
-
     @Override
     public void delete(Integer id) {
-        cartRepository.deleteById(id);
+        if (cartRepository.existsById(id)) {
+            cartRepository.deleteById(id);
+        } else throw new EntityNotFoundException("Cart not found with id: " + id);
     }
 
     @Override
-    public Cart update(Cart cart) {
-        return cartRepository.save(cart);
+    public CartDto fullUpdate(CartDto cart) {
+        if (cartRepository.existsById(cart.getId())) {
+            return mapper.mapTo(cartRepository.save(mapper.mapFrom(cart)));
+        } else throw new EntityNotFoundException("Cart not found with id: " + cart.getId());
+    }
+
+    @Override
+    public CartDto partialUpdate(CartDto cart) {
+        if (cartRepository.existsById(cart.getId())) {
+            return mapper.mapTo(cartRepository.save(mapper.mapFrom(cart)));
+        } else throw new EntityNotFoundException("Cart not found with id: " + cart.getId());
     }
 
     @Override
